Move loadBrocfile function into separate module
