+++ lib/builder.js
@@ -2,7 +2,6 @@
 
 var path = require('path')
 var fs = require('fs')
-var findup = require('findup-sync')
 var RSVP = require('rsvp')
 var tmp = require('tmp')
 var rimraf = require('rimraf')
@@ -30,7 +29,7 @@ tmp.setGracefulCleanup()
 // versions. Backwards compatibility is only guaranteed for plugins, so any
 // plugin that works with Broccoli 1.0 will work with 1.x.
 
-exports.Builder = Builder
+module.exports = Builder
 function Builder(outputNode, options) {
   if (options == null) options = {}
 
@@ -392,26 +391,6 @@ Builder.prototype.checkDirectoryExistence = function(sourcePath) {
     err.message = err.message.replace(/, stat '[^'\n]*'$/m, '')
     throw err
   }
-}
-
-
-exports.loadBrocfile = loadBrocfile
-function loadBrocfile () {
-  var brocfile = findup('Brocfile.js', {
-    nocase: true
-  })
-
-  if (brocfile == null) throw new Error('Brocfile.js not found')
-
-  var baseDir = path.dirname(brocfile)
-
-  // The chdir should perhaps live somewhere else and not be a side effect of
-  // this function, or go away entirely
-  process.chdir(baseDir)
-
-  var tree = require(brocfile)
-
-  return tree
 }
