+++ lib/cli.js
@@ -58,6 +58,6 @@ function broccoliCLI () {
 }
 
 function getBuilder () {
-  var tree = broccoli.loadBrocfile()
-  return new broccoli.Builder(tree)
+  var node = broccoli.loadBrocfile()
+  return new broccoli.Builder(node)
 }
