+++ lib/index.js
@@ -1,6 +1,5 @@
-var builder = require('./builder')
-exports.Builder = builder.Builder
-exports.loadBrocfile = builder.loadBrocfile
+exports.Builder = require('./builder')
+exports.loadBrocfile = require('./load_brocfile')
 exports.server = require('./server')
 exports.getMiddleware = require('./middleware')
 exports.Watcher = require('./watcher')
